feat: real-time import progress on queue page with 0.1% granularity
When importing from the queue, the progress bar now switches from
download progress to a green import-specific progress bar showing
the current stage (Scanning → Analyzing → Hashing → Copying) with
percentage, file counts, and the file being processed.

Worker progress reporting changed from u8 (integer %) to f32 for
0.1% granularity instead of the previous 10% jumps. Copy operations
now publish progress via an AtomicU32-backed polling task at 200ms
intervals, bridging sync copy I/O with async Redis progress events.

The queue API enriches "importing" items with live progress from
ScanResultConsumer by matching download_id between the job tracker
and the queue item.

+++ src/core/messaging/mod.rs
@@ -311,8 +311,8 @@ pub enum Message {
         files_total: usize,
         /// Number of files fully processed so far
         files_processed: usize,
-        /// Overall percent complete (0-100)
-        percent: u8,
+        /// Overall percent complete (0.0-100.0, supports 0.1% granularity)
+        percent: f32,
         /// Detail string after probe: e.g. "1080p x265 HDR10", or "unchanged" for skipped files
         detail: Option<String>,
         /// Entity IDs associated with this scan (filled by consumer, not by worker)
